Initial commit
I decided to start this project by defining a small roadmap
(versioned) of where I want to go with this project and some of
the steps required to reach there. The main goal of the indicated
roadmap is having a functional server providing a published
microservice using either JSON or XML.

To reach that goal, I found that the first step was to implement
an abstract parser for later parsing routes. The decision for an
abstract parser now is a vision that this parser will be useful
while implementing view templates in a future step (not on the
roadmap).