Created the route pattern builder
The route pattern will be responsible for matching URIs to locate
the appropriate method to handle a web request to the server. It
can be built from a resource class or from a method. Names and
parameters are identified automatically and mapped to form the
expected URI.

Some of these routes might not be exactly what the developer
expected of them and this will be handled in the next step.