Added support for customising routes in classes and methods
Given that resource classes and methods might require the use of
customised routes, the @Route annotation was created to supply
the means to provide these custom routes. Custom routes must be
used to provide length restrictions to route segments. These will
enforce that a segment must match the given length restriction to
match a route.

Also a few more tests were added to ensure the area name of
resource classes would be working as expected.