Added support for mapping methods from classes to endpoints
The RouteTable class was created to provide a mapping of paths,
HTTP methods and methods that can be used to respond to an HTTP
request. So far the table can only register resource classes to
be used but a later implementation will add behaviour to search
for the right path to answer to a request.