Added support for HTTP method annotations
By default, the engine will use the name of the method to define
the name of the HTTP method it will respond to. These annotations
allow for methods with different names to be assigned another
HTTP method.

At first this implementation was supposed to allow only one
annotation to be present in a given method but after some
consideration I decided to allow all annotations to be
registered. The decision came from the servlet implementation in
which the default implementation of the doPost method simply
calls the doGet method (thus making one method respond to two
HTTP methods).