Added support for content parameters
As of default, methods with parameters that do not implement
a static #valueOf(String) method do not get translated into a
valid endpoint. The current change enables the last parameter to
be a concrete class that does not have the valueOf method. This
parameter will be filled by trying to parse the content body of
the HTTP request into an object of this type automatically. An
error while parsing the content body shall result in the method
not even being called while not having an actual body shall
result in the method being called with this parameter set to
null.

Also was added the @NonResource annotation to enabled concrete
classes to be excluded from automatic registration of routes.