Added route resolution and HTTP error exceptions
First in this change were implemented method calls to enable the
discovery of a method handling a route through the basic parts of
the HTTP request (method and path). This implementation already
provides the means to execute correctly the matched route method
but does not handle content bodies yet.

Also were implemented a initial set of HttpExceptions. Following
the concept for the engine that the developer must worry only
with her code, using exceptions to signal an unexpected state
is coherent with OOP. However only client error codes and server
error codes will be handled in this way.