Created internal mechanism for dependency injection
The mechanism created behaves externally as a real dependency
injection mechanism but internally its behaviour is mostly that
of a service locator (thus the names used for these classes and
package). However, the service locator is not available for
external classes and thus injection can only happen where
expected on the engine.

With so many dependency injection libraries in the market, it may
be unreasonable to create another one. But the need becomes more
evident when the methods of the IServiceMap interface are
observed. Services are registered per a level of reuse and allow
for both a concrete class or a function to supply the objects
when they are needed to be created. Moreover the putSingleton
method also supports providing the actual object which will
provide the required service. This is useful to have the same
object provide two distinct services.

Also a helper method was created to ease executing methods using
the dependency injection mechanism.