Enabled the injection of services in resource constructors
No newer classes were required here. Wired the newly created
injection engine with the RouteTable to enable registered
services to be excluded from the route of the resource and have
service instances be injected in the list of arguments of the
matched route.

All tests regarding routes had to be updated since the RouteTable
is now dependent on an IServiceMap.