Parsing the content body of requests as an argument
Allowing developers access to the HTTP request would be enough
to allow them to handle the content body of those requests but it
would go against the philosophy of this framework which is to
shield developers from any direct HTTP interaction. Thus an
engine was devised to automatically parse the content body in
several known formats and transform it into an object received as
argument by the method that will handle the request.

Right now, it is possible to parse content using this class from
JSON, XML and HTML forms.

An HttpRequest class was created to enable reading the request
data into an object once and the RouteTable was adapted to allow
resolving the method to be called from an HttpRequest to enable
a richer set of information to be parsed.