Created a generic handler chain to handle received requests
The HandlerMediator class and the NextHandler class enable the
definition of a chain of responsibility in handling a request.
All arguments which can be used by handlers come from the same
service map. A handler for the end of the line has already been
implemented but may other are still missing.

Alas, the service map an the service locator classes have been
unified into a single class/interface and the methods that were
previously part of the locator were made public on the
IServiceMap interface. This will enable any class to call the
service locator given the map; I'd like to investigate the use
of a security context to prevent classes from outside the project
to actually invoke these methods.