Added support for automatic production of content for response
Most of the implementation was ready on the previous commit (thus
the last commit containing only one new test) and this fixed the
missing gaps into producing the HTTP result information from a
simple request. This new set of tests work much akin to how the
final implementation for the server will work.

A few classes were also reorganised in order to make some
subclassing and testing work properly.