Enabled custom content to be returned from methods
All of the possible method results were being handled as if
everything were an object, thus this change introduces the
possibility of returning byte arrays or input streams to be
handled as raw content of the HTTP response. To enable methods
returning these types, an annotation and a input stream wrapper
were created to enable the correct content type to be sent with
the response.

The @Route annotation was moved to the core package as that
package might contain the most common types needed to work with
in child projects.