Enabled the configuration of services
Some services might need to be configured prior to being used
but this could not be made without using instance specific
methods or classes. In order to break this dependency, the
current implementation supports using a more dependency free
approach which can be used to get options for registered
singleton services (other levels of reuse would mean the
configurations would not live longer and thus be useless).

To test the configuration strategy works, the configuration
options for routes was implemented and tested.

A few adjusts were made to the generation of XML content for
primitive types and exceptions in order not to cope with some
default or null generation.