Server is up and running
As of this commit, it is already possible to create a valid
application which will respond with the resources that exist in
the application project.

Some minor fixes were made to reflection calls and a new argument
was introduced in the Application class to enable testing of the
environment specific configurations.