Added support for controlling the number of running tasks
In a normal development environment, it should hardly be an issue
to run an unknown number of threads - most of the time, there
should be at most two or three threads running at any given time.
But during production, staging or other environments, it should
be important to control this due to limited resources on the
server running the application. This change introduced the means
to control this.

Also the test read from files for the HttpRequest class have been
adjusted to use CRLF as the parser was optimised to properly
recognise this sequence in requests.