Added command line options to applications
Given the nature of some of the options defined in the previous
commit, I decided it was more interesting if some could be set
through the command line when effectively running the application
and thus enabled some more options to be set through the same
argument list. More options may be added in future changes.

Tests were created or changed accordingly to reflect the changes
to the Application class.