Implemented logging support
With so many options for log engines available in Java, the
choice to create a new one for this project may seem
counterproductive but it does not lack reasoning.

Most engines rely on multiple dependencies in order to provide
their own log engine. By implementing one ourselves we ensure
there is no need for so many external dependencies to start
working. Also having its own log engine may foster other
developers to not waste time discussing which log engine to use
and adopt the one already provided. The log engine already
provides handlers for saving the log to the console or to a file
with the option of providing a customised template for the output
in a simple language easy to understand the information being
produced.

The created log engine also ensures a bunch of information about
the application context is immediately available in the log entry
and to be used by log handlers, dismissing the developer from
having to create wrappers to include these data or to have to
perform complex operations multiple times when registering a log
entry.

This engine must still go through a better extensive testing
before it can be validated, so we're pending identify additional
tests to perform on the engine, including performance tests
(some can be obtained through the current tests).

Also the log engine currently lacks an option for a remote log
store (most specifically syslog). That will be implemented later.