Enhanced logging services and initial log messages
Log functionality was enhanced to enable logs to be recorded in a
separated thread to reduce blocking the current request in a log
operation (log entries now go to a queue and are recorded by a
separate thread).

Tests were also written to ensure logging to console and to file
were working properly.