Adopted the servlet specification
Although the development of the engine was meant to be simple,
there was a significant interest in enabling applications using
it to be run using more robust/reliable/proven existing
application servers. Thus the need to adopt the servlet
specification.

A lot of the code had to be refactored or created in order for
this implementation to work as well as some wrapper classes and
moving some methods to other classes.

During the development of this change, the servlet implementation
was tested only using a Tomcat server since a Wildfly server
failed to run (servlet specification should be the same, right?).
Future tests with other servers may be required.