Added support for redirect results
It was a bit tricky (for me) to decide how to handle redirections
since they are neither successful operations nor
errors/exceptions but time was mostly what made this change take
so long.

Also documented the HttpInputStream class and noticed a minor
issue related to it that may be evaluated next or following it.