Added support for asynchronous tasks
Changes in this version enable publicly accessible methods in
resources to be executed asynchronously. Also it allows method to
return either a Runnable or a Callable instance to be executed
asynchronously. Results are available through a path that can be
configured during development. The queue for asynchronous
execution can be shared by all tasks of the application or have
one for asynchronous tasks separated from HTTP requests.

To support this, an entire implementation which is a competitor
to Java's default ExecutorServices was required to contain the
specific requirements of this executor pool and tests were
designed to ensure it works properly.

Also a few unused methods were removed to tidy the workspace.