Made service registration purposeful and consistent
During early stages of development, I noticed it was possible for
devs to register the same service with different reuse levels for
different environments. This change addresses a better
consistency in reuse levels in any environment by requiring the
service locator class to be annotated with its desired level.
This also restricted one can no longer register any type as a
service, especially JDK own types since they are not annotated
with the reuse level.

If some may see this as a problem, since most Java types can no
longer be readily used as services, it makes the decision to
register a service a bit more conscious, because there will be
an added overhead in creating an annotated type to register as
service and may lead devs to consider more carefully if that is
the correct answer for their problem or if it should be enhanced.

Also concerned about developer using reflection over the internal
service types, I decided to create an isolating wrapper used when
registering the core services so devs cannot use reflection or
force casting the given type into another. Thus, some of the
internal types have also been refactored to be publicly available
for other purposes and verification of internal access was
completely removed.