Implemented logging with multiple handlers and syslog support
Support for the syslog protocol has been a must since logging
was first implemented but was left out of previous refactorings
due to supposed complexity (which did not exist as can be
verified by the SyslogHandler implementation).

Also decided it was interesting to refactor a bit of the logging
to support an application having multiple handlers at once
without loss of previously supported features.

This commit has also seen a bigger change in how a Logger is
created and separated creation of objects related to the settings
file into a separate class and an enhancement in how properties
are read from such file.