Removed dependency of ConnectionInfo by loggers
One final issue with logging was that it was dependent on a class
that did not belong to its package: ConnectionInfo. This class
was best located at the core.hosting package but had to be placed
somewhere else to prevent a cyclic dependency between packages.

With this change, it is possible to allow developers to define
new variables and extra information objects to be used with
formatted log handlers (not yet implemented).