Fixed issues with null log messages and syslog IETF format
Fixed an issue where null messages could be sent for logging,
which could cause unexpected exceptions to be thrown.

Also fixed an issue with SyslogHandler messages encoded in IETF
format that the specification (RFC 5424) defined that when
messages were encoded in UTF-8 they should be prefixed with a BOM
("\u00EF\u00BB\u00BF").