Ignoring transient fields in forms and multi-part forms
I've been looking for a way to prevent hackers from tampering
with fields that should not be exposed. This is a common attack
called mass assignment. In my research, I found that whitelisting
fields (fields marked as exposed are serialisable otherwise they
are not) is more desirable than blacklisting (everything is
exposed except that which is marked as not exposed) but
supporting so in JSON and XML would require a huge amount of
work.

Thus, instead, I decided to support blacklisting as it was
already supported by default in JSON and XML, by marking a field
with the Java keywords 'transient' or 'volatile'. I decided not
to include another annotation (as done by other libraries) to
avoid introducing more annotations and push the use of the
keywords as the preferred method for preventing exposing fields
to this attack.

Also moved the package with serialisation support to the project
level (core) to reflect new organisation that is unfolding for
the project.