Added support for user identity and method protection
I believe this is one of the most important introductions in the
engine. As of now, it is possible to receive requests that
contain an identity handle which can be translated into an
identity to enable restricting the actions an authenticated or
unauthenticated user can perform on an application. User identity
is optional and if no identity provider is configured, no action
will be considered protected.

Also refactored how many of the initial configurations of
applications were made. Now most configurations may be done using
and instance of IApplicationOptions and removed the setup in two
configuration methods (now only one is necessary). This change
also made it possible to better configure the responder chain
and removed the need for services to implement any interface in
order to be further configured.

Custom services that require configurations will require the use
of a supplier to create the configured instances of the service.