Added support for validating submitted content
So far, action methods had to perform any validation of content
params received themselves and throwing any exceptions (hopefully
a BadRequestException) listing any errors from validating the
fields of that parameter object (most likely one at a time). The
current change allows objects to bear a validate method that can
evaluate and hold a list of messages to be sent back
automatically as the response of the request.

However, even if validation fails, other steps of the Responder
chain will still be evaluated, such as the identity (allowing
callers to know they cannot perform a certain operation is more
important that notifying them their object did not pass
validation).