Added support for receiving query string values into parameters
By default, the value of each parameter from an action method is
extracted from the request URL (except for the content body).
Despite being already possible to read query string parameters
using an HttpRequest object cached by the resource constructor,
this was not convenient for developers, since it would still
require them to parse and convert its parts into usable values.

In this change I introduced a simple annotation, @QueryParam,
which is used to indicate the value of a parameter of an action
method comes from a query string variable. This shortens the
steps required to use a query string parameter and removes
demands for resource classes to hold object it might not use.

Also decided to reintroduce the HeaderMap class to better group
and manage headers in requests. This time, it can only be read
once created (internally) and is not available as a service for
requests.