Added support for validation using JSR 380
The current change was one of the most challenging in this
project so far (thus the time between commits). I decided to add
support for full validation of content objects using the JSR 380
specification and use the same validation constraints to create
routes (instead of the previous implementation that required such
constraints to be declared on the route using the @Route
annotation).

In order to support this implementation, much of the
specification was not implemented/supported. The idea behind the
current implementation was to provide support for basic
validations out-of-the-box, such that no additional dependencies
were required upfront. Examples of scrapped features were ability
to add more value extractors or use expressions in error
messages.

Also, a few fixes were made to tests and code due to updating the
internal SonarQube server from v6.4 to v7.1.