Of varargs, content as InputStream and @ContentType constraint
In this change we introduced the possibility of using varargs as
a catchall for variable length paths. To introduce this support,
a few changes had to be made to how route patterns are produced
and maintained (since catchall have the least precedence over
other routes).

Also introduced the possibility of using InputStreams or, more
specifically, HttpInputStreams to receive the content body of
requests instead of only parsing from known formats. This enables
requests to simply upload files, which still might have to be
handled by the application.

Additionally to the latter, the @ContentType annotation now works
as a constraint for both return values and parameters.