Added support for localisation of resources and messages
This is an issue that went missing from JSR 380 support. But
instead of providing the same kind of support, I decided to add a
better support for localised messages enabling the list of
accepted languages sent by the client to interfere with the
result of message interpolation, that is, violation messages
using the internal validation engine are not only localised to
the locale of the application but to the locale of the client
too.

Shall JSR 380 specification or any other that may succeed it add
such support, this code is to be reviewed to add the same
support.