Added support for defining a rule to reject an async task
As stated in the documentation in this change, the default rule
for not rejecting a single request for an async task may not be
suitable for every request of an async method. This change
introduces the possibility to specify a different rule to be used
with a given method or a whole type.

We decided to add multiple support for an enumeration with the
most common rules to reject an async task but found out it could
not be enough to all cases, thus we also added support for
applications to define a class that could perform more complex
decision as to whether to reject an async task or not. We believe
the enum will be enough for most cases.

Also fixed an issue that prevented classes from being annotated
with @Async and introduced the annotation @NotAsync to indicate
a single method is not async in an async class.