Multiple bug fixes
In this commit we fixed a few minor bugs that appeared while
working on the sample project sh_rt:
- Shortened the description of the method registered to a route
  to exclude the return value (now shows signature only)
- Changed the serialiser to include JSON among accepted results
  if the client did not include it but accepts */*
- Fixed a bug that removed the wrong parameter from route parsing
  if the route receives a content object
- Added support for serialising and deserialising Java 8 dates
  represented in ISO format in JSON
- Fixed an issue where custom constraint annotations with nested
  constrains would still be required to provide a validator
- Fixed issues of alignment and remotion of expressions in some
  localised validation messages

However, this commit also left two test case failing for XML
serialisation. This was done because regular JAXB was not
producing a satisfactory result and it was decided the best would
be to implement an enhanced XML serialiser. Since this is a big
change on its own, it was decided to put it on a separate commit.