Changed control over XML support in applications
It was decided from the previous commit that a bigger change
would be necessary to address the issue of internally handle
java.time.* objects to XML (should have been addressed by Java 8
but was not). It was worked to a point in which the result
was very closely resembling how Gson worked and it was nice.

However, after further considerations, it was decided that XML
should be treated as a legacy format, that is to be supported
explicitly when required (thus the "enable.xml" setting) and that
any applications that need to support XML should do so explicitly
too.

Thus, it was decided the current implementation of XML
serialisation support was going to be maintained as is but that
developers would retain control over whether or not to enable
this support, since it should require explicit use of annotations
to control the XML result (all root serialised classes would at
least require the use of @XmlRootElement thus making it
explicit).

The created code for XML serialisation in this iteration was
saved and may be completed in a future work and thus its use may
be thought over again.