Major changes to identity and asynchronous services
This commit was supposed to be only about changing the
configuration of the identity services to receive a reference to
a class that implements IIdentityProvider instead of the actual
instance. This was done so that identity providers could use
injected services (like databases). Besides that, several other
changes were introduced:

- Changed the creation of identities to use a builder
- Ensured identities were immutable once created
- Added support for multiple @Authorize annotations
- Added a method to generate random tokens per user
- Added rules to reject repeated concurrent async calls
- Added options to config defaults rejection strategy in async
- Removed the @Resource annotation
- Updated Javadoc in several files

However, I left missing on purpose the implementation of
#hashCode() in identity instance built from using IdentityBuilder
in order to research for a best algorithm for it.