Added support for generating the "Date" HTTP header
According to RFC 7231, almost all responses to a request must
include the "Date" header. In this commit, we provided the
current date/time of the request as the default value for this
header and also provided an annotation (@DateHeader) to indicate
a method that can provide a more accurate value for the header
per request object.

Also decided to move pne level up the whole set of annotations,
exceptions, the Redirect class and the HttpInputStream class that
form the essence of the API one level up (to the
com.lmpessoa.services.core package).