Enhanced localisation support
In this change, I enhanced the localization support to use an
idiom with fewer code and more reusability of the messages. This
new idiom has been tested with the localisation of all exception
messages throughout the application.

From this change in localisation, since too many new messages
were introduced and I was unable to, alone, provide adequate
transaltions for all languages previously supported for
validation, I opted to drop support for such languages and wait
for their transaltion on the site to include back.