Initial commit of the Views module
This module is a very simple implementation that can be validated
this very few tests, as can be seen. It provides a simple
demonstration of how the core engine can be extended using
Responders.

The Views module provides renderisation of HTML pages from the
results of resource calls. These do not override the ability of
the engine to respond normally to regular REST requests (although
it could be done). The module was also implemented in order to
support with multiple template languages, although requiring a
bit more of configuration (and less standardisation) to work but
makes it very flexible.