Minor fixes
This change fixes some typos and writings in Javadoc and changes
messages from ErrorMessages to CoreMessages (thus possibly
creating a standard that message classes are prefixed with the
modules's name, as in ViewsMessage). Also fixed an issue where
catchall routes would be ignored if another route with part of
its path was already registered.