Initial commit of the Liquid render engine
This is a full implementation of the Liquid renderisation engine
(the same used by sites like Shopify) standalone and ready to be
used. It is also expansible as it allows for new tags and filters
to be registered with the engine (which will be particular to a
single application).

Currently, these classes lack any and all documentation. These
will be provided later.