Fixed login using Views engine
In order to permit external modules to use specialised responses, some
changes had to be made to the core project. Now any responder may
return an object implementing HttpResponse and it will be handled by
the upper responders as the final result and will not tamper with it
(other custom responders, however are not required to and may change
these results). This was necessary in order to allow the Views module
to send cookies to the client browser.

Additional tests were introduced to ensure the correctness of the
login process by the Views responder.